requirements.txt
autopep8==1.4.3
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
google-api-core==1.7.0
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-automl==0.1.2
google-cloud-core==0.29.1
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.18.0
httplib2==0.12.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
